Test validation better
