+++ packages/orama/tests/main.test.ts
@@ -1,30 +1,6 @@
 import t from 'tap'
-import { DocumentsStore } from '../src/components/documents-store.js'
-import { Index } from '../src/components/index.js'
-import { create, insert, insertMultiple, search } from '../src/index.js'
+import { create, insert, search } from '../src/index.js'
 import { SUPPORTED_LANGUAGES } from '../src/components/tokenizer/languages.js'
-import type { Document } from '../src/types'
-import dataset from './datasets/events.json' assert { type: 'json' }
-
-interface BaseDataEvent extends Document {
-  description: string
-  lang: string
-  category1: string
-  category2: string
-  granularity: string
-}
-
-interface DataEvent extends BaseDataEvent {
-  date: string
-}
-
-interface WrongDataEvent extends BaseDataEvent {
-  date: number
-}
-
-interface DataSet {
-  result: { events: DataEvent[] }
-}
 
 t.test('language', t => {
   t.plan(5)
@@ -108,201 +84,9 @@ t.test('language', t => {
   })
 })
 
-t.test('document validation', t => {
-  t.plan(3)
-
-  t.test('should compare the inserted doc with the schema definition', async t => {
-    t.plan(2)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-    })
-
-    t.ok(await insert(db, { quote: 'hello, world!', author: 'me' }))
-
-    await t.rejects(() => insert(db, { quote: 'hello, world!', author: true }), {
-      code: 'SCHEMA_VALIDATION_FAILURE',
-      message: 'Cannot insert document due schema validation failure on "author" property.',
-    })
-  })
-
-  t.test('should allow doc with missing schema keys to be inserted without indexing those keys', async t => {
-    t.plan(6)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-    })
-    await insert(db, {
-      quote: 'hello, world!',
-      authors: 'author should be singular',
-    })
-
-    t.equal(Object.keys((db.data.docs as DocumentsStore).docs).length, 1)
-
-    const docWithExtraKey = { quote: 'hello, world!', foo: { bar: 10 } }
-
-    const insertedInfo = await insert(db, docWithExtraKey)
-
-    t.ok(insertedInfo)
-    t.equal(Object.keys((db.data.docs as DocumentsStore).docs).length, 2)
-
-    t.ok('foo' in (db.data.docs as DocumentsStore).docs[insertedInfo]!)
-    t.same(docWithExtraKey.foo, (db.data.docs as DocumentsStore).docs[insertedInfo]!.foo)
-    t.notOk('foo' in (db.data.index as Index).indexes)
-  })
-
-  t.test(
-    'should allow doc with missing schema keys to be inserted without indexing those keys - nested schema version',
-    async t => {
-      t.plan(6)
-      const db = await create({
-        schema: {
-          quote: 'string',
-          author: {
-            name: 'string',
-            surname: 'string',
-          },
-          tag: {
-            name: 'string',
-            description: 'string',
-          },
-          isFavorite: 'boolean',
-          rating: 'number',
-        },
-      })
-      const nestedExtraKeyDoc = {
-        quote: 'So many books, so little time.',
-        author: {
-          name: 'Frank',
-          surname: 'Zappa',
-        },
-        tag: {
-          name: 'books',
-          description: 'Quotes about books',
-          unexpectedNestedProperty: 'amazing',
-        },
-        isFavorite: false,
-        rating: 5,
-        unexpectedProperty: 'wow',
-      }
-      const insertedInfo = await insert(db, nestedExtraKeyDoc)
-
-      t.ok(insertedInfo)
-      t.equal(Object.keys((db.data.docs as DocumentsStore).docs).length, 1)
-
-      t.same(
-        nestedExtraKeyDoc.unexpectedProperty,
-        (db.data.docs as DocumentsStore).docs[insertedInfo]!.unexpectedProperty,
-      )
-
-      t.same(
-        nestedExtraKeyDoc.tag.unexpectedNestedProperty,
-        ((db.data.docs as DocumentsStore).docs[insertedInfo]!.tag as unknown as Record<string, string>)
-          .unexpectedNestedProperty,
-      )
-
-      t.notOk('unexpectedProperty' in (db.data.index as Index).indexes)
-      t.notOk('tag.unexpectedProperty' in (db.data.index as Index).indexes)
-    },
-  )
-})
-
-t.test('orama', t => {
-
-  t.test('should correctly insert and retrieve data', async t => {
-    t.plan(4)
-
-    const db = await create({
-      schema: {
-        example: 'string',
-      },
-    })
-
-    const ex1Insert = await insert(db, { example: 'The quick, brown, fox' })
-    const ex1Search = await search(db, {
-      term: 'quick',
-      properties: ['example'],
-    })
-
-    t.ok(ex1Insert)
-    t.equal(ex1Search.count, 1)
-    t.type(ex1Search.elapsed.raw, 'number')
-    t.equal(ex1Search.hits[0].document.example, 'The quick, brown, fox')
-  })
-
-  t.test('should be able to insert documens with non-searchable fields', async t => {
-    t.plan(2)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-        isFavorite: 'boolean',
-        rating: 'number',
-      },
-    })
-
-    await insert(db, {
-      quote: 'Be yourself; everyone else is already taken.',
-      author: 'Oscar Wilde',
-      isFavorite: false,
-      rating: 4,
-    })
-
-    await insert(db, {
-      quote: 'So many books, so little time.',
-      author: 'Frank Zappa',
-      isFavorite: true,
-      rating: 5,
-    })
-
-    const searchResult = await search(db, {
-      term: 'frank',
-    })
-
-    t.equal(searchResult.count, 1)
-    t.equal(searchResult.hits[0].document.author, 'Frank Zappa')
-  })
-
-  t.test('should suport batch insert of documents', async t => {
-    t.plan(2)
-
-    const db = await create({
-      schema: {
-        date: 'string',
-        description: 'string',
-        lang: 'string',
-        category1: 'string',
-        category2: 'string',
-        granularity: 'string',
-      },
-    })
-
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const docs = (dataset as DataSet).result.events.slice(0, 4000)
-    const wrongSchemaDocs: WrongDataEvent[] = docs.map(doc => ({ ...doc, date: +new Date() }))
-
-    try {
-      await insertMultiple(db, docs)
-      t.equal(Object.keys((db.data.docs as DocumentsStore).docs).length, 4000)
-
-      // eslint-disable-next-line no-empty
-    } catch (_e) {}
-
-    await t.rejects(() => insertMultiple(db, wrongSchemaDocs as unknown as DataEvent[]))
-  })
-
-  t.end()
-})
-
 t.test('orama - hooks', t => {
   t.plan(2)
+
   t.test('should validate on orama creation', async t => {
     t.plan(1)
